Bump up flink versions
