+++ pom.xml
@@ -37,10 +37,10 @@
         <maven.compiler.target>1.8</maven.compiler.target>
 
         <!--Application-->
-        <flink.java.version>1.6.2</flink.java.version>
-        <flink.streaming.java_2.11.version>1.6.2</flink.streaming.java_2.11.version>
-        <flink.clients_2.11.version>1.6.2</flink.clients_2.11.version>
-        <couchbase.java.client.version>2.6.2</couchbase.java.client.version>
+        <flink.java.version>1.10.0</flink.java.version>
+        <flink.streaming.java_2.11.version>1.10.0</flink.streaming.java_2.11.version>
+        <flink.clients_2.11.version>1.10.0</flink.clients_2.11.version>
+        <couchbase.java.client.version>2.7.14</couchbase.java.client.version>
     </properties>
 
     <dependencies>
@@ -91,7 +91,7 @@
                                 <!-- add Main-Class to manifest file -->
                                 <transformer
                                         implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>com.github.isopropylcyanide.flinkcouchbasesink.job.FlinkDatabaseStartupJob</mainClass>
+                                    <mainClass>com.github.isopropylcyanide.flinkcouchbasesink.FlinkDatabaseStartupJob</mainClass>
                                 </transformer>
                             </transformers>
                             <createDependencyReducedPom>false</createDependencyReducedPom>
