+++ src/main/resources/config.properties
@@ -2,13 +2,13 @@
 couchbase.node=localhost
 
 # Couchbase username
-couchbase.username=aman
+couchbase.username=Administrator
 
 # Couchbase Password
-couchbase.password=hellohi123
+couchbase.password=password
 
 # Startup document script
-startup.documents.path=C:\\Users\\Acco-Blr-01\\Desktop\\couchbasestarter\\src\\main\\resources\\startup-documents.json
+startup.documents.path=/Users/aman.g/Work/Shenanigans/flink-couchbase-data-starter/src/main/resources/startup-documents.json
 
 # Startup document polling behaviour
 startup.documents.poll.continuous=false
